pyproject.toml
[project]
name = "lunchhunter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.2.0",
    "aiosqlite>=0.19.0",
    "geopy>=2.4.1",
    "loguru>=0.7.3",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]
